pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops"
version = "0.1.0"
description = "MLOPS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.10.1",
    "pylint>=3.3.7",
    "pynvml>=12.0.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pylint>=3.3.3",
    "pytest>=8.3.5",
    "ruff>=0.11.13",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
]
onnx = [
    "onnx==1.18.0",
    "onnxruntime-gpu>=1.20.0",
]
torch = [
    "torch>=2.7.0",
]
train = [
    "scikit-learn>=1.6.1",
    "optuna>=4.3.0",
    "matplotlib>=3.10.3",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
lint.select = ["I"]
fix = true
